requirements.txt
yfinance>=0.2.65
pandas>=1.3.0
numpy>=1.16.5
requests>=2.31.0
yahoo_fin>=0.1.8
sentence-transformers>=2.2.0
faiss-cpu>=1.11.0
openai>=1.97.1
groq>=0.30.0
beautifulsoup4>=4.11.1
pyarrow>=10.0.0
streamlit>=1.28.0
plotly>=5.0.0
scikit-learn>=1.0.0
